planCharge-76 Détecter les surcharges - Dév/évo/fonc : Amélioration de l'affichage avant/aprè période demandée.

+++ planCharge-metier/src/main/java/fr/gouv/agriculture/dal/ct/planCharge/metier/service/ChargeService.java
@@ -289,7 +289,7 @@ public class ChargeService extends AbstractService {
     }
 
     public Double chargeArrondie(@NotNull Double charge) {
-        return Math.round(charge * NBR_HEURES_OUVREES_DS_UN_JOUR) / NBR_HEURES_OUVREES_DS_UN_JOUR;
+        return (double) Math.round(charge * NBR_HEURES_OUVREES_DS_UN_JOUR) / NBR_HEURES_OUVREES_DS_UN_JOUR;
     }
 
 }
